added ipfs collective metadata update to handlePoolDetailsChanged
