+++ packages/subgraph/src/mappings/pool.ts
@@ -1,4 +1,4 @@
-import { BigInt, Bytes, ipfs, json, JSONValueKind, log } from '@graphprotocol/graph-ts';
+import { BigInt, log } from '@graphprotocol/graph-ts';
 import {
   PoolCreated,
   PoolDetailsChanged,
@@ -14,13 +14,13 @@ import {
   Claim,
   Collective,
   EventData,
-  IpfsCollective,
   PoolSettings,
   ProvableNFT,
   SafetyLimits,
   Steward,
   StewardCollective,
 } from '../../generated/schema';
+import { createOrUpdateIpfsCollective } from './ipfsCollective';
 
 export function handlePoolCreated(event: PoolCreated): void {
   const poolAddress = event.params.pool;
@@ -66,62 +66,9 @@ export function handlePoolCreated(event: PoolCreated): void {
     directPaymentPool.save();
 
     // IpfsCollective
-    let ipfsCollective = IpfsCollective.load(poolAddress.toHexString());
-    if (ipfsCollective === null) {
-      const data = fetchFromIpfsWithRetries(ipfsHash, 3);
-      if (data === null) {
-        log.error('Failed to fetch IPFS data using hash {} for collective {}', [ipfsHash, poolAddress.toHexString()]);
-        return;
-      }
-      ipfsCollective = new IpfsCollective(poolAddress.toHexString());
-      // mutates ipfsCollective
-      parseIpfsData(data, ipfsCollective, ipfsHash, poolAddress.toHexString());
-      ipfsCollective.save();
-    }
+    createOrUpdateIpfsCollective(poolAddress.toHexString(), ipfsHash);
   }
 }
-
-// mutates ipfsCollective
-function parseIpfsData(data: Bytes, ipfsCollective: IpfsCollective, ipfsHash: string, poolAddress: string): void {
-  // parse bytes to json
-  const jsonParseResult = json.try_fromBytes(data);
-  if (jsonParseResult.isError) {
-    log.error('Invalid JSON data found at IPFS hash {} for collective {}', [ipfsHash, poolAddress]);
-    return;
-  }
-  const jsonValue = jsonParseResult.value;
-
-  // make sure json is object
-  if (jsonValue.kind != JSONValueKind.OBJECT) {
-    log.error('Invalid JSON data found at IPFS hash {} for collective {}', [ipfsHash, poolAddress]);
-    return;
-  }
-  const jsonObject = jsonValue.toObject();
-
-  ipfsCollective.name = jsonObject.isSet("name") ? jsonObject.get('name')!!.toString() : "";
-  ipfsCollective.description = jsonObject.isSet("description") ? jsonObject.get('description')!!.toString() : "";
-  ipfsCollective.email = jsonObject.isSet("email") ? jsonObject.get('email')!!.toString() : null;
-  ipfsCollective.website = jsonObject.isSet("website") ? jsonObject.get('website')!!.toString() : null;
-  ipfsCollective.twitter = jsonObject.isSet("twitter") ? jsonObject.get('twitter')!!.toString() : null;
-  ipfsCollective.instagram = jsonObject.isSet("instagram") ? jsonObject.get('instagram')!!.toString() : null;
-  ipfsCollective.threads = jsonObject.isSet("threads") ? jsonObject.get('threads')!!.toString() : null;
-  ipfsCollective.headerImage = jsonObject.isSet("headerImage") ? jsonObject.get('headerImage')!!.toString() : null;
-  ipfsCollective.logo = jsonObject.isSet("logo") ? jsonObject.get('logo')!!.toString() : null;
-  ipfsCollective.images = jsonObject.isSet("images")
-    ? jsonObject.get('images')!!.toArray().map<string>((value) => value.toString())
-    : null;
-}
-
-function fetchFromIpfsWithRetries(ipfsHash: string, retries: i32): Bytes | null {
-  let data = ipfs.cat(ipfsHash);
-  let i = retries;
-  while (i > 0 && data === null) {
-    data = ipfs.cat(ipfsHash);
-    i--;
-  }
-  return data;
-}
-
 
 export function handlePoolDetailsChanged(event: PoolDetailsChanged): void {
   const poolAddress = event.params.pool;
@@ -134,6 +81,9 @@ export function handlePoolDetailsChanged(event: PoolDetailsChanged): void {
   }
   directPaymentPool.ipfs = ipfsHash;
   directPaymentPool.save();
+
+  // IpfsCollective
+  createOrUpdateIpfsCollective(poolAddress.toHexString(), ipfsHash);
 }
 
 export function handlePoolVerifiedChange(event: PoolVerifiedChanged): void {
